need to get bookId parameter to redirect to book logs after a log is deleted

+++ src/main/java/nl/marisabel/imReading/readingLogs/LogRepository.java
@@ -10,7 +10,7 @@ import javax.persistence.EntityManager;
 import java.util.List;
 
 @Repository
-public interface LogRepository extends CrudRepository<LogEntity, Long> {
+public interface LogRepository extends CrudRepository<LogEntity, Integer> {
 
 
     List<LogEntity> byBookId(@Param("bookId") BooksEntity bookId);
